+++ core/pkg/runtime/from_config.go
@@ -4,6 +4,8 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/open-feature/flagd/core/pkg/otel"
+	"go.opentelemetry.io/otel/exporters/prometheus"
 	"net/http"
 	"regexp"
 	msync "sync"
@@ -27,6 +29,7 @@ const (
 	syncProviderGrpc       = "grpc"
 	syncProviderKubernetes = "kubernetes"
 	syncProviderHTTP       = "http"
+	svcName                = "openfeature/flagd"
 )
 
 var (
@@ -52,10 +55,16 @@ func FromConfig(logger *logger.Logger, config Config) (*Runtime, error) {
 		sources = append(sources, sync.URI)
 	}
 	s.FlagSources = sources
+	exporter, err := prometheus.New()
+	if err != nil {
+		return nil, err
+	}
 	rt := Runtime{
-		config:    config,
-		Logger:    logger.WithFields(zap.String("component", "runtime")),
-		Evaluator: eval.NewJSONEvaluator(logger, s),
+		config:      config,
+		Logger:      logger.WithFields(zap.String("component", "runtime")),
+		Evaluator:   eval.NewJSONEvaluator(logger, s),
+		metrics:     otel.NewOTelRecorder(exporter, svcName),
+		serviceName: svcName,
 	}
 	if err := rt.setSyncImplFromConfig(logger); err != nil {
 		return nil, err
@@ -75,6 +84,7 @@ func (r *Runtime) setService(logger *logger.Logger) {
 		Logger: logger.WithFields(
 			zap.String("component", "service"),
 		),
+		Metrics: r.metrics,
 	}
 }
