+++ core/pkg/service/iservice.go
@@ -25,6 +25,7 @@ type Configuration struct {
 	ReadinessProbe ReadinessProbe
 	Port           uint16
 	MetricsPort    uint16
+	ServiceName    string
 }
 
 /*
